Backend for token validator
