+++ eox_tenant/edxapp_wrapper/auth.py
@@ -23,12 +23,3 @@ def get_edx_auth_failed():
     backend = import_module(backend_function)
 
     return backend.get_edx_auth_failed()
-
-
-def get_edx_oauth2_validator_class():
-    """ Backend to get EdxOAuth2Validator class
-    """
-    backend_function = settings.EOX_TENANT_EDX_AUTH_BACKEND
-    backend = import_module(backend_function)
-
-    return backend.get_edx_oauth2_validator_class()
