+++ eox_tenant/edxapp_wrapper/backends/edx_auth_i_v1.py
@@ -2,9 +2,6 @@
 from openedx.core.djangoapps.oauth_dispatch.dot_overrides.backends import (  # pylint: disable=import-error
     EdxRateLimitedAllowAllUsersModelBackend,
 )
-from openedx.core.djangoapps.oauth_dispatch.dot_overrides.validators import (  # pylint: disable=import-error
-    EdxOAuth2Validator,
-)
 from openedx.core.djangoapps.user_authn.exceptions import AuthFailedError  # pylint: disable=import-error
 
 
@@ -16,9 +13,3 @@ def get_edx_auth_backend():
 def get_edx_auth_failed():
     """ Backend to get the AuthFailedError class. """
     return AuthFailedError
-
-
-def get_edx_oauth2_validator_class():
-    """ Backend to get EdxOAuth2Validator class
-    """
-    return EdxOAuth2Validator
